[+] Небольшие доработки
