+++ blog/templates/base.html
@@ -13,7 +13,7 @@
 <body>
 {% include 'navbar.html' %}
 <div class="container-fluid">
-  {% block body %}{% endblock %}
+    {% block body %}{% endblock %}
 </div>
 
 {% block js %}
