+++ blog/templates/navbar.html
@@ -14,9 +14,12 @@
                     <a class="nav-link" href="{{ url_for('articles.articles_list') }}">Аrticles</a>
                 </li>
                 {% if current_user.is_authenticated %}
-                    <li class="nav-item">
-                        <a class="nav-link" href="{{ url_for('auth.logout') }}">Logout {{ current_user.email }}</a>
-                    </li>
+                    <a class="nav-link" href="{{ url_for('auth.logout') }}">Logout</a>
+                {% else %}
+                    <a href="{{ url_for('auth.login') }}"
+                       class="nav-link {% if request.endpoint == 'auth.login' -%}active{%- endif %}">
+                        Login
+                    </a>
                 {% endif %}
             </ul>
         </div>
